chore: create ryot_asset_loading crate
