+++ crates/ryot/src/bevy_ryot/drawing/commands/mod.rs
@@ -14,7 +14,7 @@
 use crate::bevy_ryot::drawing::DrawingBundle;
 use crate::bevy_ryot::{GameObjectBundle, GameObjectId};
 use bevy::prelude::*;
-use ryot_assets::prelude::FrameGroup;
+use ryot_content::prelude::FrameGroup;
 use ryot_tiled::prelude::*;
 
 mod update;
