--- crates/ryot_assets/src/core/visual_elements/mod.rs
@@ -1,4 +1,0 @@
-pub mod definitions;
-
-#[cfg(feature = "bevy")]
-pub mod load;
