+++ crates/ryot_content/src/tests.rs
@@ -1,6 +1,7 @@
-use crate::prelude::{ContentType, SpriteLayout, SpriteSheetData, SpriteSheetDataSet};
+use crate::prelude::ContentType;
 use glam::UVec2;
 use rstest::{fixture, rstest};
+use ryot_sprites::prelude::*;
 use serde_json::{from_str, to_string};
 
 #[rstest]
