+++ crates/ryot_internal/src/lib.rs
@@ -2,11 +2,19 @@
 pub mod prelude;
 
 pub mod assets {
-    pub use ryot_assets::*;
+    pub use ryot_asset_loading::*;
+}
+
+pub mod content {
+    pub use ryot_content::*;
 }
 
 pub mod core {
     pub use ryot_core::*;
+}
+
+pub mod sprites {
+    pub use ryot_sprites::*;
 }
 
 pub mod tiled {
@@ -18,9 +26,9 @@ pub mod pathfinder {
     pub use ryot_pathfinder::*;
 }
 
-#[cfg(feature = "ryot_tibia_assets")]
+#[cfg(feature = "ryot_tibia_content")]
 pub mod tibia {
-    pub use ryot_tibia_assets::*;
+    pub use ryot_tibia_content::*;
 }
 
 #[cfg(feature = "ryot_ray_casting")]
