+++ crates/ryot_internal/src/prelude.rs
@@ -1,9 +1,11 @@
-pub use ryot_assets::prelude::*;
+pub use ryot_asset_loading::prelude::*;
+pub use ryot_content::prelude::*;
 pub use ryot_core::prelude::*;
+pub use ryot_sprites::prelude::*;
 pub use ryot_tiled::prelude::*;
 
-#[cfg(feature = "ryot_tibia_assets")]
-pub use ryot_tibia_assets::prelude as tibia;
+#[cfg(feature = "ryot_tibia_content")]
+pub use ryot_tibia_content::prelude as tibia;
 
 #[cfg(feature = "ryot_pathfinder")]
 pub use ryot_pathfinder::prelude::*;
