+++ crates/ryot_sprites/Cargo.toml
@@ -1,0 +1,74 @@
+[package]
+name = "ryot_sprites"
+version = "0.2.0"
+edition = "2021"
+authors = [
+    "Lucas Grossi <lucas.ggrosis@gmail.com>",
+    "Luan Santos <cargo@luan.sh>",
+]
+license = "MIT OR Apache-2.0"
+description = "A crate to handle sprites for the Ryot game engine."
+homepage = "https://github.com/opentibiabr/Ryot"
+repository = "https://github.com/opentibiabr/Ryot"
+documentation = "https://docs.rs/ryot/"
+keywords = ["gamedev", "ryot", "sprites", "bevy"]
+categories = ["game-development", "game-engines", "games"]
+
+[package.metadata.docs.rs]
+rustdoc-args = ["-Zunstable-options", "--cfg", "docsrs"]
+all-features = true
+
+[features]
+default = ["bevy"]
+
+bevy = [
+    "dep:bevy_asset",
+    "dep:bevy_ecs",
+    "dep:bevy_render",
+    "dep:bevy_sprite",
+    "dep:bevy_utils",
+    "dep:bevy_asset_loader",
+]
+
+[dependencies.bevy_asset]
+version = "0.13"
+optional = true
+
+[dependencies.bevy_ecs]
+version = "0.13"
+features = ["bevy_reflect"]
+optional = true
+
+[dependencies.bevy_render]
+version = "0.13"
+optional = true
+
+[dependencies.bevy_sprite]
+version = "0.13"
+optional = true
+
+[dependencies.bevy_utils]
+version = "0.13"
+optional = true
+
+[dependencies.bevy_asset_loader]
+version = "0.20.1"
+optional = true
+
+[dependencies]
+ryot_core = { path = "../ryot_core", version = "0.2.0" }
+async-std = "1.12.0"
+derive_more.workspace = true
+glam.workspace = true
+serde_repr = "0.1"
+serde.workspace = true
+serde_json.workspace = true
+strum.workspace = true
+thiserror.workspace = true
+
+[dev-dependencies]
+quickcheck.workspace = true
+quickcheck_macros.workspace = true
+rstest.workspace = true
+time-test.workspace = true
+rand.workspace = true
