+++ crates/ryot_tibia_content/src/appearances.proto
