+++ crates/ryot_tibia_content/src/conversions.rs
@@ -32,7 +32,8 @@
 //! in one module, we ensure consistent behavior and data transformations throughout the application,
 //! facilitating easier updates and maintenance when dealing with external asset changes.
 use crate as tibia;
-use ryot_assets::prelude::*;
+use ryot_content::prelude::*;
+use ryot_sprites::prelude::*;
 
 impl From<tibia::VisualElements> for VisualElements {
     fn from(item: tibia::VisualElements) -> Self {
