+++ crates/ryot_tiled/src/grid.rs
@@ -7,7 +7,7 @@ use bevy_render::render_asset::RenderAssetUsages;
 use bevy_transform::components::Transform;
 use bevy_utils::default;
 use glam::Vec2;
-use ryot_assets::prelude::{SpriteLayout, TextureAtlasLayouts};
+use ryot_sprites::prelude::{SpriteLayout, TextureAtlasLayouts};
 
 pub static GRID_LAYER: Layer = Layer::Hud(0);
